# contracts/openapi/finance-api-v1.yaml
openapi: 3.1.0
info:
  title: Finance Core API
  version: 1.0.0
  description: Public API shared by React and Android. Java Core owns all financial mutations.
servers:
  - url: http://localhost:8080
    description: Local development only
security:
  - keycloakOidc: []
paths:
...
  /internal/v1/telegram/report:
    post:
      operationId: getTelegramFinanceReport
      security:
        - telegramServiceToken: []
      description: Returns the shared finance report DTO after Core revalidates the opaque actor context and report.read permission. The request cannot select a tenant or user.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TelegramReportRequest' }
      responses:
        '200':
          description: Finance report computed by Core for the actor's current tenant and selected scope
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FinanceReport' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { description: Telegram service credential is not configured }
  /internal/v1/telegram/products/catalog:
    post:
      operationId: getTelegramProductCatalog
      security:
        - telegramServiceToken: []
      description: Returns the current actor's private confirmed-purchase catalog after Core revalidates receipt.read. The request cannot select a user or tenant.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TelegramProductCatalogRequest' }
      responses:
        '200':
          description: Member-scoped catalog or search
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductCatalogResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { description: Telegram service credential is not configured or product history is unavailable }
...
  /internal/v1/telegram/actions/personal-inflation:
    post:
      operationId: getTelegramPersonalInflation
      security:
        - telegramServiceToken: []
      description: Returns personal receipt-price inflation for the verified Telegram actor; request cannot select a user or tenant.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResolveTelegramActorContext' }
      responses:
        '200':
          description: Member-scoped 90-day basket; insufficient history has null totals
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PersonalInflation' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { description: Telegram or price analytics service is unavailable }
  /internal/v1/telegram/actions/recurring:
    post:
      operationId: getTelegramRecurringProjection
      security:
        - telegramServiceToken: []
      description: Returns recurring series for the verified Telegram actor and active tenant. Request cannot select another user or timezone.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResolveTelegramActorContext' }
      responses:
        '200':
          description: Member-scoped recurring income, expense, due-soon, and overdue series
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecurringProjection' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { description: Telegram or recurring analytics service is unavailable }
...
  /internal/v1/telegram/summary:
    post:
      operationId: getTelegramDashboardSummary
      security:
        - telegramServiceToken: []
      description: Loads the dashboard only after Core revalidates the opaque actor context and its report.read permission. The body does not select a user or tenant.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResolveTelegramActorContext' }
      responses:
        '200':
          description: Current member's dashboard scoped to the selected tenant
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DashboardSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { description: Telegram service credential is not configured }
  /internal/v1/telegram/debts:
    post:
      operationId: listTelegramOpenDebts
      security:
        - telegramServiceToken: []
      description: Returns open debts in the current actor tenant after Core revalidates debt.read. The request does not select a member or tenant.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResolveTelegramActorContext' }
      responses:
        '200':
          description: Current tenant's open debts for an explicit debt-payment choice
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TelegramDebtList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { description: Telegram service credential is not configured }
  /internal/v1/telegram/transactions:
    post:
      operationId: listTelegramRecentTransactions
      security:
        - telegramServiceToken: []
      description: Returns the selected member's latest posted operations, ordered by Core creation time, after revalidating transaction.read. No tenant or user selector is accepted.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TelegramHistoryRequest' }
      responses:
        '200':
          description: Recent operations scoped to the current actor
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TelegramHistoryResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { description: Telegram service credential is not configured }
  /internal/v1/telegram/transactions/latest/void:
    post:
      operationId: voidTelegramLatestTransaction
      security:
        - telegramServiceToken: []
      description: Voids only the current actor's latest posted operation after revalidating transaction.write.own. A stale history selection fails, and idempotency replays the original result.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TelegramVoidLatestRequest' }
      responses:
        '200':
          description: Voided transaction; debt payment effects are reversed by Core
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transaction' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '503': { description: Telegram service credential is not configured }
  /internal/v1/telegram/transaction-drafts/repeat:
    post:
      operationId: repeatTelegramExpenseDraft
      security:
        - telegramServiceToken: []
      description: Creates a durable review draft from the actor's own posted expense without calling AI. Repeated requests with the same key return the same draft; no transaction is posted before confirmation.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TelegramRepeatRequest' }
      responses:
        '201':
          description: Repeat draft created with copied expense fields and current date
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransactionDraft' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { description: Telegram service credential is not configured }
  /internal/v1/telegram/transaction-drafts:
    post:
      operationId: createTelegramTransactionDraft
      security:
        - telegramServiceToken: []
      description: Creates a durable review draft after Core revalidates transaction.write.own. The actor token selects the active tenant and member; caller-supplied userId or tenantId fields are ignored.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTelegramTransactionDraft' }
      responses:
        '201':
          description: Review draft created; no transaction is posted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransactionDraft' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '502': { description: AI output failed validation }
        '503': { description: Telegram or AI service credential is not configured }
  /internal/v1/telegram/transaction-drafts/{draftId}/confirm:
    parameters:
      - name: draftId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: confirmTelegramTransactionDraft
      security:
        - telegramServiceToken: []
      description: Confirms only the owner's pending draft in the actor context tenant after revalidating transaction.write.own. Version and idempotency key prevent stale or duplicate financial writes.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TelegramDraftDecision' }
      responses:
        '200':
          description: Confirmed transaction
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transaction' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '422': { description: Debt payment needs an explicitly selected debt }
        '503': { description: Telegram service credential is not configured }
  /internal/v1/telegram/transaction-drafts/{draftId}/amount:
    parameters:
      - name: draftId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: updateTelegramTransactionDraftAmount
      security:
        - telegramServiceToken: []
      description: Updates only the amount of the actor's pending draft after revalidating transaction.write.own. The server owns all other fields and checks the draft version.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TelegramDraftAmount' }
      responses:
        '200':
          description: Updated review draft with incremented version
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransactionDraft' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '503': { description: Telegram service credential is not configured }
  /internal/v1/telegram/transaction-drafts/{draftId}/read:
    parameters:
      - name: draftId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: getTelegramTransactionDraft
      security:
        - telegramServiceToken: []
      description: Reads only the actor's own draft in the selected tenant after revalidating transaction.write.own. Allows callbacks to resume after Redis FSM expiry.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResolveTelegramActorContext' }
      responses:
        '200':
          description: Current tenant-scoped draft
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransactionDraft' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { description: Telegram service credential is not configured }
  /internal/v1/telegram/transaction-drafts/{draftId}/edit:
    parameters:
      - name: draftId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: editTelegramTransactionDraft
      security:
        - telegramServiceToken: []
      description: Updates owner-selected draft fields after revalidating transaction.write.own. Core scopes by actor tenant and owner and checks the current version; caller-supplied userId or tenantId fields are ignored.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TelegramDraftEdit' }
      responses:
        '200':
          description: Updated review draft with incremented version
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransactionDraft' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '503': { description: Telegram service credential is not configured }
  /internal/v1/telegram/transaction-drafts/{draftId}/cancel:
    parameters:
      - name: draftId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: cancelTelegramTransactionDraft
      security:
        - telegramServiceToken: []
      description: Cancels only the owner's pending draft in the actor context tenant after revalidating transaction.write.own and checking its version.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TelegramDraftCancel' }
      responses:
        '200':
          description: Draft cancellation result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TelegramDraftCancelled' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '503': { description: Telegram service credential is not configured }
...
  /api/v1/tenants/{tenantId}/budget-proposals:
    parameters:
      - $ref: '#/components/parameters/TenantId'
    post:
      operationId: createBudgetProposal
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BudgetProposalRequest' }
      responses:
        '201':
          description: Pending proposal; no budget changes until explicit apply
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BudgetProposal' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { description: Role cannot propose family budgets }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/tenants/{tenantId}/budget-proposals/{proposalId}/apply:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - name: proposalId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: applyBudgetProposal
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: All family limits applied atomically after explicit user action
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BudgetOverview' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { description: Role cannot apply family budgets }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '410': { description: Proposal expired }
  /api/v1/tenants/{tenantId}/budget-proposals/history:
    parameters:
      - $ref: '#/components/parameters/TenantId'
    post:
      operationId: createHistoryBudgetProposal
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201':
          description: AI proposal based on allowlisted aggregates; remains pending until explicit apply
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BudgetProposal' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { description: Role cannot propose family budgets }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { description: Planned income or 30 days of posted expense history is missing }
        '502': { description: AI response failed schema or policy validation }
        '503': { description: AI service is unavailable or not configured }
...
  /api/v1/tenants/{tenantId}/budgets/personal-overrides:
    parameters:
      - $ref: '#/components/parameters/TenantId'
    delete:
      operationId: resetPersonalBudgets
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Personal overrides removed; family limits become effective again
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BudgetOverview' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { description: Role cannot change this budget scope }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bffSession:
      type: apiKey
      in: cookie
      name: JSESSIONID
    keycloakOidc:
      type: openIdConnect
      openIdConnectUrl: https://auth.example.invalid/realms/finance/.well-known/openid-configuration
    telegramServiceToken:
      type: apiKey
      in: header
      name: X-Finance-Service-Token
...
  responses:
    BadRequest:
      description: Invalid amount, type, category or timestamp
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing or invalid OIDC identity
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Current tenant role does not permit this financial mutation
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    ResolveTelegramActorContext:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token: { type: string, minLength: 43, maxLength: 43, pattern: '^[A-Za-z0-9_-]{43}$' }
    TelegramReportRequest:
      type: object
      additionalProperties: true
      required: [token, period, scope]
      properties:
        token: { type: string, minLength: 43, maxLength: 43, pattern: '^[A-Za-z0-9_-]{43}$' }
        period: { type: string, enum: [month, week, 90d, custom] }
        month: { type: [string, 'null'], pattern: '^\d{4}-(0[1-9]|1[0-2])$' }
        from: { type: [string, 'null'], format: date }
        to: { type: [string, 'null'], format: date }
        scope: { type: string, enum: [personal, family] }
      description: Actor token determines identity and tenant. Extra caller identity or tenant fields are ignored.
    TelegramProductCatalogRequest:
      type: object
      additionalProperties: false
      required: [token, query]
      properties:
        token: { type: string, minLength: 43, maxLength: 43, pattern: '^[A-Za-z0-9_-]{43}$' }
        query: { type: string, maxLength: 80, default: '' }
    TelegramHistoryRequest:
      type: object
      additionalProperties: false
      required: [token, limit]
      properties:
        token: { type: string, minLength: 43, maxLength: 43, pattern: '^[A-Za-z0-9_-]{43}$' }
        limit: { type: integer, minimum: 1, maximum: 20, default: 8 }
...
    TelegramDebtList:
      type: object
      additionalProperties: false
      required: [debts]
      properties:
        debts:
          type: array
          items: { $ref: '#/components/schemas/Debt' }
...